Modify sub classes of mission
MissionConqueteContinent and MissionElimination used too complexe type
for entity. Database need simple value instead of complexe type.
Add Set of Joueur in estRemplie()'s arguments, it represents the list of
all player in the game.
MissionConqueteContinent got two String which represents continents.
MissionElimination got a String which is the color of a player instead
of a player.